requirements.txt
matplotlib==3.3.3
numpy==1.19.2
pandas==0.23.4
torch==1.9.0
pyyaml==5.4.1
scikit-learn==0.24.2
scipy==1.5.2
torchaudio==0.9.0
torchvision>=0.2.2
tornado==6.0.4
tqdm==4.55.1

